+++ tests/Fixtures/Models/Person.php
@@ -24,5 +24,5 @@ class Person extends Model
     public function family()
     {
         return $this->belongsTo(Family::class);
-    }     
+    }       
 }