+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# 1.6.0
+
+- Adding in static collections feature
+- Static collections are a way to hold collections of classes and other items.
+
+# 1.5.3
+
+- Validating parameter methods
+- Snake casing relation whereHasIn query
+- Snake/camel casing param query functions
+
+# 1.5.0
+
+- Adding default param query methods
+
+# v1.4.0
+
+- Adding whereHas to contact and DbRepo
+
 # v1.3.0
 
 - Adding repository restore functionality